core: Add minimal Entity, Component storages, Systems API

+++ include/engine/time.h
@@ -14,23 +14,23 @@ public:
 
 	void Update();
 
-	double GetDeltaSeconds() const { return deltaSeconds_; }
-	double GetElapsedSeconds() const { return elapsedSeconds_; }
-	double GetTimeScale() const { return timeScale_; }
+	float GetDeltaSeconds() const { return deltaSeconds_; }
+	float GetElapsedSeconds() const { return elapsedSeconds_; }
+	float GetTimeScale() const { return timeScale_; }
 	bool IsPaused() const { return isPaused_; }
 
-	void SetScale(double scale) { timeScale_ = scale; }
+	void SetScale(float scale) { timeScale_ = scale; }
 	void SetPaused(bool isPaused) { isPaused_ = isPaused; }
 
 private:
 	// Measured time
 	Clock::time_point lastTick_;
 	// Frame delta
-	double deltaSeconds_ = 0.0;
+	float deltaSeconds_ = 0.0;
 	// Total simulation time
-	double elapsedSeconds_ = 0.0;
+	float elapsedSeconds_ = 0.0;
 	// Fast-forward / slow-mo
-	double timeScale_ = 1.0;
+	float timeScale_ = 1.0;
 	// For returning zero-delta without stopping clock
 	bool isPaused_ = false;
 };
